paginate lists in commands to fix disconnect
turns out there's a limit on how long a chat message packet can be - oops!
if you did "/simplyhotsprings biomeslist all" on the dedicated server, the client would get disconnected if there are around ~130 biomes or so. this can be accomplished with just BOP!
so now a few commands are split into pages, similar to /help.